requirements.txt
langchain==0.3.2
langchain-community==0.3.1
langchain-aws==0.2.1
chromadb==0.5.11
fastapi==0.115.0
uvicorn==0.30.6
pypdf==5.0.0
python-dotenv==1.0.1
boto3==1.35.24
ragas==0.2.3
streamlit==1.38.0
yfinance==0.2.49